requirements.txt
Django==1.8
django-angular==0.7.14
django-braces==1.4.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
django-filter==0.10.0
djangorestframework==3.1.2
gunicorn==19.3.0
psycopg2==2.6.1
requests==2.6.0
six==1.9.0
sqlparse==0.1.15
